+++ packages/components/src/Modal/Modal.rebuilt.test.tsx
@@ -26,11 +26,55 @@ describe("Composable Modal", () => {
     expect(header).toBeDefined();
     expect(header).toHaveTextContent("Modal Title");
     expect(screen.getByText("This is some extra content")).toBeInTheDocument();
-    expect(screen.getByText("Submit")).toBeInTheDocument();
-    expect(screen.getByText("Cancel")).toBeInTheDocument();
-    expect(screen.getByText("Delete")).toBeInTheDocument();
+    expect(screen.getByRole("button", { name: "Submit" })).toBeInTheDocument();
+    expect(screen.getByRole("button", { name: "Cancel" })).toBeInTheDocument();
+    expect(screen.getByRole("button", { name: "Delete" })).toBeInTheDocument();
+    expect(screen.getByRole("button", { name: "Submit" })).toHaveClass(
+      "primary",
+    );
+    expect(screen.getByRole("button", { name: "Cancel" })).toHaveClass(
+      "primary subtle",
+    );
+    expect(screen.getByRole("button", { name: "Delete" })).toHaveClass(
+      "secondary destructive",
+    );
   });
 
+  it("should allow overriding of action buttons", () => {
+    render(
+      <Modal.Provider open={true}>
+        <Modal.Content>
+          <Modal.Header title="Modal Title" />
+          <Modal.Actions
+            primary={{
+              label: "Submit",
+              variation: "destructive",
+              type: "secondary",
+            }}
+            secondary={{
+              label: "Cancel",
+              variation: "work",
+              type: "tertiary",
+            }}
+            tertiary={{
+              label: "Delete",
+              variation: "destructive",
+              type: "secondary",
+            }}
+          />
+        </Modal.Content>
+      </Modal.Provider>,
+    );
+    expect(screen.getByRole("button", { name: "Submit" })).toHaveClass(
+      "secondary destructive",
+    );
+    expect(screen.getByRole("button", { name: "Cancel" })).toHaveClass(
+      "tertiary work",
+    );
+    expect(screen.getByRole("button", { name: "Delete" })).toHaveClass(
+      "secondary destructive",
+    );
+  });
   it('modal contains aria role of "dialog"', async () => {
     render(
       <Modal.Provider open={true}>